Add airbnb eslint with some configurations changes, add prettier as well to follow this config changes, rename file to jsx as linter suggested
