+++ .prettierrc.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+  trailingComma: 'es5',
+  semi: true,
+  singleQuote: true,
+};
