+++ src/styles/globals.css
@@ -1,24 +1,24 @@
 * {
-	box-sizing: border-box;
-	padding: 0;
-	margin: 0;
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
 }
 
 html,
 body {
-	max-width: 100vw;
-	overflow-x: hidden;
-	font-family: 'Courier New', Courier, monospace;
-	font-size: 20px;
+  max-width: 100vw;
+  overflow-x: hidden;
+  font-family: 'Courier New', Courier, monospace;
+  font-size: 20px;
 }
 
 a {
-	color: inherit;
-	text-decoration: none;
+  color: inherit;
+  text-decoration: none;
 }
 
 @media (prefers-color-scheme: white) {
-	html {
-		color-scheme: white;
-	}
+  html {
+    color-scheme: white;
+  }
 }
